[Refactoring] Add testscript header file.
